refactoring user's related classes
* moved mail classes
* moved presenter class
* removed enums
* removed unecessary accessors and mutators
