+++ app/Mail/Admin/ResetPassword.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Services\Auth\Front\Mail;
+namespace App\Mail\Admin;
 
-use App\Models\User;
+use App\Models\Administrator;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailable;
@@ -12,7 +12,7 @@ class ResetPassword extends Mailable implements ShouldQueue
 {
     use Queueable, SerializesModels;
 
-    /** @var \App\Models\User */
+    /** @var \App\Models\Administrator */
     public $user;
 
     /** @var string */
@@ -20,8 +20,11 @@ class ResetPassword extends Mailable implements ShouldQueue
 
     /**
      * Create a new message instance.
+     *
+     * @param \App\Models\Administrator $user
+     * @param string                    $token
      */
-    public function __construct(User $user, string $token)
+    public function __construct(Administrator $user, string $token)
     {
         $this->user = $user;
 
@@ -36,7 +39,7 @@ class ResetPassword extends Mailable implements ShouldQueue
     public function build()
     {
         return $this
-            ->subject('Toegang tot '.config('app.url'))
-            ->markdown($this->user->hasNeverLoggedIn() ? 'mails.member.setPassword' : 'mails.member.resetPassword');
+            ->subject('🔐 Toegang tot '.config('app.url'))
+            ->markdown($this->user->hasNeverLoggedIn() ? 'mails.admin.setPassword' : 'mails.admin.resetPassword');
     }
 }
