+++ app/Models/Administrator.php
@@ -2,16 +2,14 @@
 
 namespace App\Models;
 
-use App\Services\Auth\Back\Enums\UserRole;
-use App\Services\Auth\Back\Enums\UserStatus;
-use App\Services\Auth\Back\Exceptions\UserIsAlreadyActivated;
-use App\Services\Auth\Back\Mail\ResetPassword;
+use App\Mail\Admin\ResetPassword;
 use App\Services\Auth\User as BaseUser;
+use App\Services\Auth\Back\Exceptions\UserIsAlreadyActivated;
 use Illuminate\Support\Facades\Mail;
 
 /**
- * @property \App\Services\Auth\Back\Enums\UserRole $role
- * @property \App\Services\Auth\Back\Enums\UserStatus $status
+ * @property string $role
+ * @property string $status
  */
 class Administrator extends BaseUser
 {
@@ -32,16 +30,6 @@ class Administrator extends BaseUser
         return action('Back\AdministratorsController@edit', $this->id);
     }
 
-    public function getStatusAttribute(): UserStatus
-    {
-        return new UserStatus($this->attributes['status']);
-    }
-
-    public function setStatusAttribute(string $status)
-    {
-        $this->attributes['status'] = $status;
-    }
-
     public function hasStatus(string $status): bool
     {
         return $this->status === $status;
@@ -49,33 +37,18 @@ class Administrator extends BaseUser
 
     public function isActive(): bool
     {
-        return $this->hasStatus(UserStatus::ACTIVE);
+        return $this->hasStatus(BaseUser::STATUS_ACTIVE);
     }
 
     public function activate(): self
     {
-        if ($this->status !== UserStatus::WAITING_FOR_APPROVAL) {
+        if ($this->status !== BaseUser::STATUS_WAITING_FOR_APPROVAL) {
             throw new UserIsAlreadyActivated();
         }
 
-        $this->status = UserStatus::ACTIVE;
+        $this->status = BaseUser::STATUS_ACTIVE;
 
         return $this;
-    }
-
-    public function getRoleAttribute(): string
-    {
-        return $this->attributes['role'];
-    }
-
-    public function setRoleAttribute(string $role)
-    {
-        $this->attributes['role'] = $role;
-    }
-
-    public function hasRole(UserRole $role): bool
-    {
-        return $this->role === $role;
     }
 
     /**
